--- buttons.h
@@ -1,13 +1,0 @@
-#ifndef BUTTONS_H
-#define BUTTONS_H
-
-#define view_btn 2
-#define edit_btn 3
-#define up_btn 4
-#define down_btn 5
-#define start_btn 6
-
-
-void poll_buttons();
-
-#endif
