--- state_machine.h
@@ -1,16 +1,0 @@
-#ifndef STATE_MACHINE_H
-#define STATE_MACHINE_H 
-
-#include "global.h"
-
-void update_state_machine();
-
-void startup();
-void display_mash()
-void display_ferm();
-void edit_mash();
-void edit_ferm();
-void pause();
-void shut_down();
-
-#endif
